prevent supabase module from being imported in test
- extract supabase connection file to own directory
- extract supabase-involved handlers into own controller file
- conditionally import that controller file if env is not testing
